Make filter_versions a classmethod of DatabaseManager
