+++ database/__init__.py
@@ -13,7 +13,7 @@ from postgrest.base_request_builder import APIResponse
 from supabase._async.client import create_client, AsyncClient
 from bot.config import DEV_MODE
 from database.schema import VersionRecord
-from database.utils import get_version_string, filter_versions
+from database.utils import get_version_string
 
 
 class DatabaseManager:
@@ -88,11 +88,71 @@ class DatabaseManager:
         versions = cls.get_versions_list(edition=edition)[-1]
         return get_version_string(versions, no_edition=True)
 
+    @classmethod
+    def filter_versions(cls, spec: str) -> list[str]:
+        """Return all versions that match the version specification."""
+
+        def parse_version(version_str: str):
+            major, minor, patch = version_str.split('.')
+            return int(major), int(minor), int(patch)
+
+        all_versions = cls.get_versions_list("Java")
+        # Convert each version in all_versions into a tuple for easy comparison
+        all_version_tuples = [(v["major_version"], v["minor_version"], v["patch_number"]) for v in all_versions]
+
+        # Split the spec by commas: e.g. "1.14 - 1.16.1, 1.17, 1.19+"
+        parts = [part.strip() for part in spec.split(',')]
+
+        valid_tuples: list[tuple[int, int, int]] = []
+
+        for part in parts:
+            # Case 1: range like "1.14 - 1.16.1"
+            if '-' in part:
+                start_str, end_str = [p.strip() for p in part.split('-')]
+                start_tuple = parse_version(start_str) if start_str.count('.') == 2 else parse_version(start_str + '.0')
+                end_tuple = parse_version(end_str) if end_str.count('.') == 2 else parse_version(end_str + '.0')
+
+                for v_tuple in all_version_tuples:
+                    if start_tuple <= v_tuple <= end_tuple:
+                        valid_tuples.append(v_tuple)
+
+            # Case 2: trailing plus like "1.19+"
+            elif part.endswith('+'):
+                base_str = part[:-1].strip()
+                # If user just wrote "1.19+", assume "1.19.0"
+                if base_str.count('.') == 1:
+                    base_str += '.0'
+                base_tuple = parse_version(base_str)
+
+                for v_tuple in all_version_tuples:
+                    if v_tuple >= base_tuple:
+                        valid_tuples.append(v_tuple)
+
+            # Case 3: exact version or prefix, e.g. "1.17" or "1.17.1"
+            else:
+                subparts = part.split('.')
+                # If only major.minor specified (like "1.17"), match all "1.17.x"
+                if len(subparts) == 2:
+                    major, minor = map(int, subparts)
+                    for v_tuple in all_version_tuples:
+                        if v_tuple[0] == major and v_tuple[1] == minor:
+                            valid_tuples.append(v_tuple)
+                # If a full version specified (like "1.17.1"), match exactly that version
+                elif len(subparts) == 3:
+                    v_tuple = tuple(map(int, subparts))
+                    if v_tuple in all_version_tuples:
+                        valid_tuples.append(v_tuple)
+                else:
+                    # Optionally, handle malformed inputs or major-only specs
+                    pass
+
+        return [f"{major}.{minor}.{patch}" for major, minor, patch in valid_tuples]
+
 
 async def main():
     await DatabaseManager.setup()
     spec_string = "1.14 - 1.16.1, 1.17, 1.19+"
-    print(await filter_versions(spec_string))
+    print(DatabaseManager.filter_versions(spec_string))
 
 if __name__ == "__main__":
     import asyncio
